combine NotebooksIndex container and display
